scripting
git-svn-id: https://svn2.hosted-projects.com/jjinux/jenarix@81 af7818e5-df21-49d5-b9e6-ecf3e1c3d4f7

+++ sandbox/warren/090514/src/test/test002.c
@@ -217,7 +217,7 @@ int main(int argc, char **argv)
 
     {
       jx_ob str = jx_ob_from_str("1");
-      P1("1 == %d", jx_ob_as_int( jx_ob_to_int(str)));
+      P1("1 == %d", (int)jx_ob_as_int( jx_ob_to_int(str)));
       jx_ob_free(str);
     }
